refactor to save duplication
